feat(vm): add i16/i32 checked integer semantics

+++ tests/unit/test_il_parse_invalid_type.cpp
@@ -12,7 +12,7 @@
 int main()
 {
     const char *src = R"(il 0.1.2
-extern @foo(i32) -> i64
+extern @foo(i128) -> i64
 )";
     std::istringstream in(src);
     il::core::Module m;
